requirements.txt
numpy
matplotlib
scipy
scikit-image
Pillow
noise
